Rez and death animation

+++ Assets/Scripts/Healthbar.cs
@@ -7,7 +7,7 @@ public class Healthbar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetHealth(int currentHealth, int maxHealth)
+    public void SetHealth(float currentHealth, float maxHealth)
     {
         slider.value = currentHealth;
         slider.maxValue = maxHealth;
